030-security-monitoring/oauth-msal-starter/requirements.txt
Flask
Flask-Login
Flask-Session
msal
python-dotenv